kafka/examples/requirements.txt
confluent-kafka
requests
uvicorn[standard]
sseclient-py
pyspark
fastapi
folium
psycopg2-binary